requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pinecone-client==5.0.1
sentence-transformers==3.0.1
groq==0.9.0
pydantic==2.12.5
typing_extensions==4.15.
discord.py==2.6.4